Don't respond to stop signal until all resources are used.
So selectWhileNotStopped just stops the select loop but doesn't respond
to the stop signal. It now takes a ReceiveChan<Unit>, as it's not
concerned with the response.

launchStoppable, on the other hand, now doesn't wait for a response to
a stop signal; the response is just the launched goroutine finishing,
either as a response to the stop signal or not.

So now the select loops on the subscribe consumer just stop early, and
the stop() call doesn't return until the consumer returns, closing
resources on its way.