Stop consuming on requeue.
This way, we don't require the underlying queue system to react to a
requeue, either explicit (Requeue is returned by the handler) or
implicit (crash/timeout).

This is handy for us because Postgres doesn't seem to provide a way to
hook into transaction failure, so it won't resend notifications if we
don't ack the requeue, ie. in a crash mode. So we move the
responsibility to the consumer: if you couldn't OK the message, you have
to restart the consumer.