Add aborted itersess package, for the record.
I realized I needed to implement cancellation in cspiter, but adding it
with the given implementation would be brittle. So I thought I could
reimplement it in a more type-driven style inspired by session types.

Then I realized I was fixated to my early decision to make producers
take a "send handle" an block until done. It's way simpler to have a
coroutine-like API, and with a bit of extra boilerplate it can be
made to comply with structured concurrency.

This is being done in the upcoming coro package.